requirements.txt
discord.py
aiohttp
python-dotenv
requests
PyNaCl